+++ src/subscribers/thanks.ts
@@ -54,7 +54,7 @@ export class ThanksSubscriber extends Subscriber {
         title: pr.title,
       });
 
-      await context.octokit.rest.issues.createComment(context.issue({ body }));
+      await this.notice(context, pr.number, body);
 
       log.info(`Commented on PR #${pr.number}`);
     });